improved widget routing
