--- public/.idea/libraries/Flutter_for_Android.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="Flutter for Android">
-    <CLASSES>
-      <root url="jar:///Users/David/flutter/bin/cache/artifacts/engine/android-arm/flutter.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
