--- public/.idea/modules.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ProjectModuleManager">
-    <modules>
-      <module fileurl="file://$PROJECT_DIR$/public.iml" filepath="$PROJECT_DIR$/public.iml" />
-      <module fileurl="file://$PROJECT_DIR$/android/public_android.iml" filepath="$PROJECT_DIR$/android/public_android.iml" />
-    </modules>
-  </component>
-</project>
